Implementation of array functions every and splice.
I spent a lot of time considering how I might have a template string in
JS like `My dog is ${trait}` such that the trait is shown in bold while
the other text remains plain. Scott showed me how to use the
createElement function on the document object, and we tried to style the
created elements that way, but it was to complicated and bloated. We
then realized that the easy way to make part of the sentence bold was to
abort the mission of making part of the template string bold and to
instead go to the html code and add spans.

Lesson learned: If something seems like a weird challenge, go back to
the HTML and see if there is a simple way to code it using tags.